example-projects/domain-check/pyproject.toml
[tool.poetry]
name = "domain-check"
version = "0.1.0"
description = "A1"
authors = ["Donatas Rasiukevicius <donatas@fly.io>"]
readme = "README.md"
packages = [{include = "domain_check", from = "src"}]
include = [ "templates" ]

[tool.poetry.dependencies]
python="^3.11"
fastapi=">=0.108.0,<0.109.0"
uvicorn=">=0.34.2,<0.35.0"
python-whois=">=0.9.5,<0.10.0"
jinja2=">=3.1.6,<4.0.0"
python-multipart=">=0.0.20,<0.0.21"
dnspython = "^2.7.0"
pyopenssl = "^25.0.0"
requests = "^2.32.3"
redis = "^5.0.1"
dotenv = "^0.9.9"
pydantic = {extras = ["email"], version = "^2.11.4"}
apscheduler = "^3.11.0"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


